diskinstaller: Add syslinux support, rebase on HC
This is a port of the GB syslinux changes to diskinstaller, rebased
on upstream HC diskinstaller. It replaces uses of GRUB with Syslinux.

Changes from the GB version:

- The system image is now loopback mounted from the data partition
in the live environment; there's no need to have another copy of it
in a partition.
- live_img target removed; there is now only installer_img and it comes
with a live menu option by default.
- Default syslinux menu option is to boot from harddisk for absent-minded
individuals who leave their USB key in after installing

Change-Id: I22398bd0c2758bac0e4665c221ba6e25a660ef6e