diskinstaller: Use squashfs to store images
Using squashfs reduces the total size of all partitions to about 109
MB, from over 500. Since squashfs partitions are read-only and the
live environment does not need to be persistent, a tmpfs is used
for the live environment's /data partition.

A 10MB writable 'stash' partition exists as a convenience when copying
log and other small files off the device during debugging.

Change-Id: I6e988db49065559d76de192f143046fdf86f9c72