editdisklbl: Don't set dinfo->num_lba
This confuses libdiskconfig, isn't necessary with the changes to
how regular files are handled, and the calculations were wrong
anyway (EBR records weren't accounted for).

The Makefile rules were changed such that the destination file passed
to libdiskconfig is of zero length, so that either a concrete value
in disk_layout.conf, or an automatically computed value is used (based
on the sum of the partition sizes) to determine the total disk size.

Change-Id: Iddfec43ea46c128a52547d796388b52b0c49ab8a
Signed-off-by: Andrew Boie <andrew.p.boie@intel.com>
Depends-Change-Id: Ie11b2aafccd864a98b753bfcc3459bd94df591c2