HACK: installer: Zero out space bewteen MBR and first ptn
Some PC devices will enter a recovery mode specified in the
Windows Boot Manager if the device is powered on with a magic key.
When this happens and Android is installed, the device becomes
unbootable until reinstalled.

This mechanism seems to rely on some metadata in unpartitioned space
after the MBR but before the first partition. Clearing this space
causes the magic shortcut to have no effect.

This is not a root-cause fix and it would be better to disable
the Windows Boot Manager entirely, but technical documentation
on how it works and where the relevant bits lie is currently
unknown.

Change-Id: Ic45f90cf653ab327e7dc685b7f91c4743332b14e
Signed-off-by: Andrew Boie <andrew.p.boie@intel.com>