Enable diskinstaller to process sparse ext4 images
Android is able to produce a sparse user data image which contains
only data but not free space. It reduces the size of the image.
However, it is an Android-only feature so we need to write a proper
ext4 image onto the block device. This patch uses the simg2img
utility to expand the sparse image and write the data onto
the block device.

For: AXIA-939
Change-Id: I36bd044e9004019d1891257e4bc8cf703917c109
Depends-Change-Id: I2b2bf08636087125c431db79c7cdabc1ba22609b
Signed-off-by: Daniel Leung <daniel.leung@intel.com>