Installer feature to support footer parameter
Added support for installer config file to support a new type of
parameter: 'footer'. This new parameter will allow resizing of a file
system but leaving a required amount of free space at the end of the
partition. This amount is specified in footer parameter.
The syntax of the parameter is:
footer <size>[K|M|G]
where <size> is the number of bytes for the footer free space.
The size can be preceded by a suffix:
- K: the size is in kilobytes
- M: the size is in megabytes
- G: the size is in gigabytes.

For: AXIA-946
Change-Id: I0c548cdf70114538669cb7c6b387a2854a23a5eb
Signed-off-by: Robert Chiras <robert.chiras@intel.com>