config.mk: only try to copy stuff in system partition
The Makefile is attempting to assemble a list of output artifacts
associated with a set of modules. However, some modules have the same
LOCAL_MODULE for both their host and target versions. The Makefile
tries to do a patsubst of $(PRODUCT_OUT) with the installer ramdisk,
but since these host files don't have $(PRODUCT_OUT) in their paths,
no substitution is done and it essentially calls $(ACP) with the source
and destination being the same path. $(ACP) helpfully truncates the
file to 0-length in this case instead of reporting an error.

This fixes it by filtering out anything in $(android_sysbase_files) that
isn't in $(PRODUCT_OUT).

Issue: AXIA-1038
Change-Id: I2ae7654c0077d0dbcf398044da4744bd065d482e
Signed-off-by: Andrew Boie <andrew.p.boie@intel.com>